Unfinished: document plan for webAO's virtual base
Basically, there's more compatibility problems trying to
make the server send asset IDs than if one were simply to let the
client figure out all of the compatibility problems on its own.
One way to do this would be to mount a "virtual base" that
essentially maps folders to asset packages, and if the file is not
found, then it falls back to HTTP sources. (This means that named
paths within a package will be allowed - hashes will not be required
to refer to an asset within a package.)